tests: TDD affiliated dealer usecase

+++ internal/controllers/affiliated_dealers/domain/domain.go
@@ -3,8 +3,8 @@ package domain
 import "context"
 
 type Usecase interface {
-	Create(ctx context.Context, arg *CreateRequestParams) (int, error)
-	Update(ctx context.Context, id int, arg *UpdateRequestParams) error
-	FindByID(ctx context.Context, id int) (*AffiliatedDealer, error)
+	Create(ctx context.Context, arg *CreateRequestParams) (int64, error)
+	Update(ctx context.Context, id int64, arg *UpdateRequestParams) error
+	FindByID(ctx context.Context, id int64) (*AffiliatedDealer, error)
 	List(ctx context.Context, arg *FetchParams) (*AffiliatedDealerList, error)
 }
